requirements.txt
numpy==1.26.4
torch==2.8.0
psutil==7.0.0
sympy==1.14.0
jinja2==3.1.6
markupsafe==3.0.2
networkx==3.4.2
filelock==3.19.1
fsspec==2025.9.0
typing-extensions==4.15.0
gymnasium==0.29.1
opencv-python
matplotlib
imageio
crafter==1.8.1
stable-baselines3[extra]==2.1.0
tqdm
tensorboard